refactor: Extract sk_methods.rs
